Added Auto Identity Map Extension (<1G), Fixed Issues with String stdlib, Updates to VMM Helper Routines

+++ tools/kernelrtl/strings.cpp
@@ -32,7 +32,7 @@ i32 KernelRTL::strncmp(const char* s1, const char* s2, usize len)
 }
 
 void* KernelRTL::memset(void* dest, u64 value, u64 len) {
-    u8* ptr = (u8*) dest;
+    u64* ptr = (u64*) dest;
     while (len-- > 0) {
         *(ptr++) = value;
     }
